+++ bin/modules/UlsOutput.py
@@ -532,6 +532,12 @@ class UlsOutput:
         #aka_log.log.info(f"{self.name} new HTTP output string: {self.http_out_format} ")
         return True
 
+    def ingest_os_vars_into_output_format(self):
+        self.tcpudp_out_format = os.path.expandvars(self.tcpudp_out_format)
+        self.http_out_format = os.path.expandvars(self.http_out_format)
+
+        return True
+
     ### / FEATURE REQ 20240318 - https://github.com/akamai/uls/issues/57
 
 # EOF